Company_AI_webiste_builder/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
openai==0.28.1